add webservconfig.cpp; add config file to command arg
