+++ default.conf
@@ -12,7 +12,7 @@ server:
 		autoindex: on
 
 	location /askme:
-		rewrite: 302 http://www.google.com
+		rewrite: 302 http://codam.nl
 			
 	error_page: 404 ./default_pages/404.html
 	error_page: 500 ./default_pages/404.html
